+++ StaffDatabaseFunctions.h
@@ -92,14 +92,14 @@ void setStaffInfo(/*bool canSetStaffInfo,*/ int staffIndex, int fieldIndex, stri
 	//recreate staff data file
 	ofstream output;
 	output.open("Staff_Data.txt");
-	for (int i = 0; i < v1.size(); i++)//put all of v1 back into the staff data file
+	for (int i = 0; i < static_cast<int>(v1.size()); i++)//put all of v1 back into the staff data file
 	{
 		temp = v1.at(i);
 		output << temp;
 	}
 	output << newData;//put the new info into the file
 	output << ' ';
-	for (int i = 0; i < v2.size(); i++)//put all of v2 back into the end staff data file
+	for (int i = 0; i < static_cast<int>(v2.size()); i++)//put all of v2 back into the end staff data file
 	{
 		temp = v2.at(i);
 		output << temp;
@@ -138,7 +138,7 @@ void createTextFile()
 
 	ofstream output;
 	output.open("Staff_Data.txt");
-	for (int i = 0; i < v1.size(); i++)//put vectors into file
+	for (int i = 0; i < static_cast<int>(v1.size()); i++)//put vectors into file
 	{
 		temp = v1.at(i);
 		output << temp;
@@ -187,7 +187,7 @@ void createUsernameFile()
 
 	ofstream output;
 	output.open("Staff_Usernames.txt");
-	for (int i = 0; i < v1.size(); i++)//put into file rotating vectors
+	for (int i = 0; i < static_cast<int>(v1.size()); i++)//put into file rotating vectors
 	{
 		temp = v1.at(i);
 		output << temp;
